requirements.txt
streamlit
openai
langchain
nltk
scikit-learn
SpeechRecognition
python-dotenv
markdown
bs4
